move alert.py from root to alert
