+++ alert/webhook.py
@@ -4,7 +4,7 @@ import json
 import httplib
 from urlparse import urlparse
 
-from alert_util import match_alert
+from alert_util import match_alert, AlertException
 
 
 class Webhook(object):
@@ -29,11 +29,14 @@ class Webhook(object):
 
         if match_alert(config['routes'], level):
             url_info = urlparse(url)
-
             port = 80 if url_info.port is None else url_info.port
-            http_client = httplib.HTTPConnection(url_info.hostname, port, timeout=5)
-
-            http_client.request("POST", url_info.path, json.dumps(params), headers)
+            try:
+                http_client = httplib.HTTPConnection(url_info.hostname,
+                                                     port, timeout=5)
+                http_client.request("POST", url_info.path,
+                                    json.dumps(params), headers)
+            except Exception as e:
+                raise AlertException(e)
 
     @staticmethod
     def check_config(config):
